dynamic generation of number of TSV files, and run tests in separate directories (#125)
* support running tests in separate temp directories

* find scm when run in different working directory for version

* dynamically generate TSV files using XSLT based on XML output, rather than hardcoding

* for haplotypes or LD for 3 or more loci, handle all file generation, dynamically
(including loci greater than 4)

* use pytest fixtures globally across tests

* delete temporary test directories by default, keep if verbosity flag is set

* refactoring pytest boilerplate to check lists of files and check that only expected are generated, extra or unexpected files, or missing files for TSV tests will cause test to fail

* new test cases: 5-locus test case for longer haplos and test with both 3 and 4 locus haplotypes

* disable 5 locus for Windows case - some numerical issues

+++ src/PyPop/Meta.py
@@ -324,8 +324,12 @@ class Meta:
                 if len(fileBatchList) > 1:
                     for dat in datfiles:
                         if success:
-                            #print("renaming:", dat, "to:", "%s.%d" % (dat, fileBatch))
-                            os.rename(dat, "%s.%d" % (dat, fileBatch))
+                            if os.path.exists(dat):
+                                #print("renaming:", dat, "to:", "%s.%d" % (dat, fileBatch))
+                                os.rename(dat, "%s.%d" % (dat, fileBatch))
+                            else:
+                                print("%s in batch %d doesn't exist - skipping"
+                                  % (dat, fileBatch))
                         else:
                             print("problem with generating %s in batch %d"
                                   % (dat, fileBatch))
@@ -340,6 +344,8 @@ class Meta:
                 for i in range(len(fileBatchList)):
                     # write temp file to outdat
                     catFilename = "%s.%d" % (dat, i)
+                    if not(os.path.exists(catFilename)):  # if the file is not generated, we skip
+                        continue
                     catFile = open(catFilename)
                     # drop the first line, iff we are past first file
                     for linenum, line in enumerate(catFile):
@@ -351,6 +357,9 @@ class Meta:
                     # then remove it
                     catFile.close()
                     os.remove(catFilename)
-                # finally, close the ultimate output file
+                # close the ultimate output file
                 outdat.close()
+                # if the file is empty, we remove it
+                if os.path.getsize(dat) == 0:
+                    os.remove(dat)
